+++ razomy/nuxt/src/module.ts
@@ -6,7 +6,7 @@ export interface ModuleOptions {
 
 export default defineNuxtModule<ModuleOptions>({
   meta: {
-    name: '@razomy/vue-nuxt',
+    name: '@razomy/nuxt',
     // configKey: 'myModule',
   },
   // Default configuration options of the Nuxt module